Replace table by css flexbox grid nasa-gibs/worldview#123
